chest_x-ray/requirements.txt
Flask==3.0.3
flask-cors==4.0.0
torch==2.2.2
numpy==1.26.4
Pillow==10.3.0
opencv-python-headless==4.10.0.84
requests==2.32.3
gunicorn==23.0.0
